notifications menu overhaul complete
